+++ demo/basic-store/package.json
